Revise variational framework.
We now define e.g., the whitened variational Gaussian process as $q(f(\cdot))$ as `gpx.WhitenedVariationalGaussian(prior=prior, inducing_inputs=z)` and can directly make predictions on this object.
